tests/wpt/web-platform-tests/referrer-policy/css-integration/font-face/internal-stylesheet.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS integration - Font from internal stylesheet</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/utils.js"></script>
    <!-- Common global functions for referrer-policy tests. -->
    <script src="/referrer-policy/generic/common.js"></script>
    <meta name="referrer" content="origin">
  </head>
  <body>
    <p>Check that resources from internal stylesheets are loaded with
    the referrer and referrer policy from the document.</p>

    <div class="styled"></div>

    <script>
      let css_test = async_test("Font from internal stylesheet.");
      let id = token();
      let css_url = location.protocol + "//www1." + location.hostname + ":" +
                    location.port +
                    "/referrer-policy/generic/subresource/font.py" + "?id=" +
                    id + "&type=font";
      let font_url = css_url + "&report-headers";

      let style = document.createElement("style");
      style.textContent = "@font-face { font-family: 'wpt'; font-style: normal; font-weight: normal; src: url(" + css_url + "); format('truetype'); } body { font-family: 'wpt';}";
      document.head.appendChild(style);
      css_test.step_timeout(
          queryXhr.bind(this, font_url,
                        function(message) {
                          css_test.step(function() {
                            assert_own_property(message, "headers");
                            assert_own_property(message, "referrer");
                            assert_equals(message.referrer, location.origin + "/");
                          });
                          css_test.done();
                        }),
          1000);
      </script>

    <div id="log"></div>
  </body>
</html>
